remove rake task and fix reindex task
The fill_db and picky_index task was removed because it took to long. This task needs to be done on the host directly. The reindex task was broken because the filename was wrong and not in the root folder. Maybe this one also needs to be removed and run directly on the host.